feat:swal-before-delete

+++ src/components/CardWatchlist.jsx
@@ -4,12 +4,31 @@ import { AiOutlinePlus } from "react-icons/ai";
 import { MdDelete } from "react-icons/md";
 import { deleteRepo } from '../store/action/actionCreator'
 import { useDispatch } from 'react-redux';
+import Swal from 'sweetalert2';
+
 
 const CardWatchlist = ({ repo }) => {
    const dispatch = useDispatch();
 
    const handleDelete = () => {
-      dispatch(deleteRepo(repo._id));
+      Swal.fire({
+         title: 'Are you sure?',
+         text: "You won't be able to revert this!",
+         icon: 'warning',
+         showCancelButton: true,
+         confirmButtonColor: '#3085d6',
+         cancelButtonColor: '#d33',
+         confirmButtonText: 'Yes, delete it!'
+       }).then((result) => {
+         if (result.isConfirmed) {
+           Swal.fire(
+             'Deleted!',
+             'Your file has been deleted.',
+             'success'
+           )
+           dispatch(deleteRepo(repo._id));
+         }
+       })
    };
    return (
       <div id="card" className="bg-white border p-5 rounded-md relative group hover:bg-slate-100">
